jpql native query -> queryDsl 변경 작업 중

+++ banchanshop/src/main/java/com/bbsk/banchanshop/admin/dto/ResponseRecently7DaysOrderCount.java
@@ -11,5 +11,5 @@ import java.time.LocalDate;
 public class ResponseRecently7DaysOrderCount {
 
     private String day;
-    private int orderCount;
+    private Long orderCount;
 }
